requirements.txt
streamlit==1.26.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
langchain==0.0.208
streamlit-chat==0.0.2.2
python-dotenv==1.0.0
